INDEX_UI/requirements.txt
flask
pandas
mysql-connector-python
matplotlib
seaborn
folium
